src/main/java/com/oocl/cultivation/ParkingBoy.java
package com.oocl.cultivation;

public class ParkingBoy {
    private ParkingLot parkingLot;

    public ParkingBoy(){
        parkingLot = new ParkingLot();
    }

    public ParkingBoy(ParkingLot parkingLot){
        this.parkingLot = parkingLot;
    }

    public ParkingTicket park(Car car) {
        return parkingLot.isFull() ? null : parkingLot.park(car);
    }

    public Car fetchCar() {
        return null;
    }

    public Car fetchCar(ParkingTicket parkingTicket) {
        return parkingLot.fetchCar(parkingTicket);
    }
}
